--- a/src/fisher_scoring/__init__.py
+++ b/src/fisher_scoring/__init__.py
@@ -13,3 +13,10 @@ __all__ = [
     "FisherScoringMultinomialRegression",
     "FisherScoringFocalRegression",
 ]
+
+# Add dynamic version retrieval
+try:
+    from importlib.metadata import version
+    __version__ = version("fisher-scoring")
+except ImportError:
+    __version__ = "unknown"