… .DS_Store
… dist/.DS_Store
… dist/fisher_scoring-0.1.4-py3-none-any.whl
… dist/fisher_scoring-0.1.4.tar.gz
… dist/fisher_scoring-2.0-py3-none-any.whl
… dist/fisher_scoring-2.0.tar.gz
… docs/.DS_Store
… docs/images/title.png
… poetry.lock
… requirements.txt
… src/.DS_Store
… src/fisher_scoring/.DS_Store
… src/fisher_scoring/__pycache__/fisher_scoring_focal.cpython-310.pyc
… src/fisher_scoring/__pycache__/fisher_scoring_logistic.cpython-310.pyc
… src/fisher_scoring/__pycache__/fisher_scoring_multinomial.cpython-310.pyc
… tests/__pycache__/test_fisher_scoring_focal.cpython-310-pytest-8.3.3.pyc
… tests/__pycache__/test_fisher_scoring_logistic.cpython-310-pytest-8.3.3.pyc
… tests/__pycache__/test_fisher_scoring_multinomial.cpython-310-pytest-8.3.3.pyc