## v2.0

### Performance Improvements
Optimized matrix calculations for substantial speed and memory efficiency across all models. This version introduces streamlined operations, delivering up to 290x faster convergence. Detailed performance gains by model:

- **Multinomial Logistic Regression**: Training time reduced from 125.10s to 0.43s (~290x speedup).
- **Logistic Regression**: Training time reduced from 0.24s to 0.05s (~5x speedup).
- **Focal Loss Logistic Regression**: Training time reduced from 0.26s to 0.01s (~26x speedup).

### Bug Fixes
- **Verbose Parameter**: The `verbose` parameter in Focal Loss Logistic Regression now functions as expected, enabling accurate logging during training.